Move private functions from QueryBuilder to a separate Private module
