+++ postgrest.cabal
@@ -39,6 +39,7 @@ library
                       PostgREST.RangeQuery
                       PostgREST.Types
   other-modules:      Paths_postgrest
+                      PostgREST.QueryBuilder.Private
   hs-source-dirs:     src
   build-depends:      base                      >= 4.9 && < 4.13
                     , HTTP                      >= 4000.3.7 && < 4000.4
